+++ azure-monitor-opentelemetry/azure/monitor/opentelemetry/_configure.py
@@ -22,9 +22,7 @@ from azure.monitor.opentelemetry.exporter import (
 )
 from azure.monitor.opentelemetry.util.configurations import _get_configurations
 from opentelemetry._logs import get_logger_provider, set_logger_provider
-from opentelemetry.instrumentation.dependencies import (
-    get_dist_dependency_conflicts,
-)
+from opentelemetry.instrumentation.dependencies import get_dependency_conflicts
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.metrics import set_meter_provider
 from opentelemetry.sdk._logs import LoggerProvider, LoggingHandler
@@ -39,15 +37,15 @@ from pkg_resources import iter_entry_points
 _logger = getLogger(__name__)
 
 
-_SUPPORTED_INSTRUMENTED_LIBRARIES = (
-    "django",
-    "fastapi",
-    "flask",
-    "psycopg2",
-    "requests",
-    "urllib",
-    "urllib3",
-)
+_SUPPORTED_INSTRUMENTED_LIBRARIES_TO_INSTRUMENTS_MAP = {
+    "django": ("django >= 1.10",),
+    "fastapi": ("fastapi ~= 0.58",),
+    "flask": ("flask >= 1.0, < 3.0",),
+    "psycopg2": ("psycopg2 >= 2.7.3.1",),
+    "requests": ("requests ~= 2.0",),
+    "urllib": tuple(),
+    "urllib3": ("urllib3 >= 1.0.0, < 2.0.0",),
+}
 
 
 def configure_azure_monitor(**kwargs) -> None:
@@ -125,13 +123,19 @@ def _setup_metrics(configurations: Dict[str, ConfigurationValue]):
 
 def _setup_instrumentations():
     # use pkg_resources for now until https://github.com/open-telemetry/opentelemetry-python/pull/3168 is merged
-    for entry_point in iter_entry_points("opentelemetry_instrumentor"):
+    for entry_point in iter_entry_points("azure_monitor_opentelemetry_instrumentor"):
         lib_name = entry_point.name
-        if lib_name not in _SUPPORTED_INSTRUMENTED_LIBRARIES:
+        if (
+            lib_name
+            not in _SUPPORTED_INSTRUMENTED_LIBRARIES_TO_INSTRUMENTS_MAP
+        ):
             continue
         try:
             # Check if dependent libraries/version are installed
-            conflict = get_dist_dependency_conflicts(entry_point.dist)
+            instruments = _SUPPORTED_INSTRUMENTED_LIBRARIES_TO_INSTRUMENTS_MAP[
+                lib_name
+            ]
+            conflict = get_dependency_conflicts(instruments)
             if conflict:
                 _logger.debug(
                     "Skipping instrumentation %s: %s",